+++ packages/terra-compact-interactive-list/src/utils/keyHandlerUtils.js
@@ -1,4 +1,4 @@
-import getFocusableElements from '../../../terra-table/src/utils/focusManagement';
+import getFocusableElements from 'terra-table/lib/utils/focusManagement';
 
 /**
  * @param {HTMLElement} element - The element to check if it is a text input