+++ packages/terra-framework-docs/src/terra-dev-site/doc/data-grid/FlowsheetDataGrid.4/Examples.6/FlowsheetWithSections.jsx
@@ -154,6 +154,7 @@ const FlowsheetWithSections = () => {
       id="flowsheet-with-sections"
       columns={gridDataJSON.cols}
       sections={tableSections}
+      ariaLabel="Flowsheet Data Grid with Sections"
       onSectionSelect={handleSectionSelect}
       onCellRangeSelect={handleCellRangeSelection}
       onCellSelect={onCellSelect}