+++ packages/terra-framework-docs/src/terra-dev-site/doc/compact-interactive-list/Examples.2/FixedWidthColumns.jsx
@@ -83,7 +83,7 @@ const cols = [
 
 const FixedWidthColumns = () => (
   <CompactInteractiveList
-    id="interactive-compact-columns-width"
+    id="compact-interactive-list-fixed-width"
     ariaLabel="Compact Interactive List"
     rows={rows}
     columns={cols}