+++ packages/terra-framework-docs/CHANGELOG.md
@@ -5,6 +5,9 @@
 * Changed
   * Updated the `terra-date-time-picker` example for Error label.
 
+* Added
+  * Added the 'aria-label' attribute to announce the information for the 'Flowsheet-Data Grid'.
+
 ## 1.62.0 - (January 26, 2024)
 
 * Added