+++ packages/terra-framework-docs/CHANGELOG.md
@@ -3,6 +3,10 @@
 ## Unreleased
 
 * Added
+  * Added documentation for exported constants of `terra-table` in the About page.
+  * Added tests for resizable `terra-table`.
+
+* Added
   * Added the 'aria-label' attribute to announce the information for the 'Flowsheet-Data Grid'.
 
 ## 1.62.0 - (January 26, 2024)